pyproject.toml
[project]
name = "nyxmon"
version = "0.1.0"
description = "Minimal monitoring package"
readme = "README.md"
authors = [
    { name = "Jochen Wersdörfer", email = "jochen-nyxmon@wersdoerfer.de" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
]

[project.scripts]
start-agent = "nyxmon.entrypoints.cli:start_agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "pytest-anyio",
]

[tool.pytest.ini_options]
addopts = "-ra"
